Remove neuron model definition
